requirements.txt
customtkinter
opencv-python
opencv-python-headless
ultralytics
pillow